+++ src/constants.js
@@ -1,0 +1,5 @@
+export const AI_WRAPPED_KEY = Symbol("AI_WRAPPED_KEY");
+
+export const MIDDLEWARE = "Middleware";
+export const PARAM_MIDDLEWARE = "Param Middleware";
+export const ERROR_MIDDLEWARE = "Error Middleware";