+++ src/isFunction.js
@@ -1,0 +1,3 @@
+const isFunction = (functionLike) => typeof functionLike === "function";
+
+export default isFunction;